Fix linting issues.
