+++ games/oblivion_remaster/constants.py
@@ -1,0 +1,1 @@
+PLUGIN_NAME = "Oblivion Remastered Support Plugin"
