+++ pyproject.toml
@@ -47,6 +47,9 @@ target-version = "py310"
 [tool.ruff.lint]
 extend-select = ["B", "Q", "I"]
 
+[tool.ruff.lint.flake8-bugbear]
+extend-immutable-calls = ["PyQt6.QtCore.QModelIndex"]
+
 [tool.ruff.lint.isort]
 known-first-party = ["mobase"]
